refactor: change domain module to pure kotlin
