+++ data/src/main/java/com/repository/MovieRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.repository
 
-import com.MovieRepository
+import com.android.domain.MovieRepository
 import com.utils.BaseRepository
 
 class MovieRepositoryImpl :MovieRepository, BaseRepository() {
