+++ domain/src/main/java/com/android/domain/MovieRepository.kt
@@ -1,4 +1,4 @@
-package com
+package com.android.domain
 
 interface MovieRepository {
 }
