+++ domain/src/main/java/com/android/domain/exception/Exception.kt
@@ -1,3 +1,3 @@
-package com.exception
+package com.android.domain.exception
 
 sealed class CineVerseException : Exception()
