+++ domain/src/main/java/com/android/domain/model/MovieActor.kt
@@ -1,4 +1,4 @@
-package com.model
+package com.android.domain.model
 
 data class MovieActor(
     val movieId: Int,
