+++ domain/src/main/java/com/android/domain/model/SeriesActor.kt
@@ -1,4 +1,4 @@
-package com.model
+package com.android.domain.model
 
 data class SeriesActor(
     val seriesId: Int,
