+++ domain/src/main/java/com/android/domain/model/SeriesSeason.kt
@@ -1,0 +1,14 @@
+package com.android.domain.model
+
+import kotlinx.datetime.LocalDate
+
+data class SeriesSeason(
+    val id: Int,
+    val number: Int,
+    val seriesId: Int,
+    val episodesCount: Int,
+    val rating: Float,
+    val description: String,
+    val releaseDate: LocalDate,
+    val poster: String,
+)
