+++ presentation/build.gradle.kts
@@ -75,4 +75,5 @@ dependencies {
     implementation(libs.koin.core)
     implementation(libs.koin.android)
     implementation(libs.koin.compose)
+    implementation(libs.kotlinx.datetime)
 }
